SDM croupier
fixed moist air without
refactor ICMW*/example.py -> ICMW*/example.py, ICMW*/simulation.py
CREATE ICMW*/example_toss_pair_benchmark
