+++ PySDM_tests/smoke_tests/ICMW_2012_case_1/test_velocity_field.py
@@ -2,7 +2,7 @@ import numpy as np
 from matplotlib import pyplot
 import pytest
 from PySDM_examples.ICMW_2012_case_1.setup import Setup
-from PySDM_examples.ICMW_2012_case_1.example import Simulation
+from PySDM_examples.ICMW_2012_case_1.simulation import Simulation
 from PySDM_examples.ICMW_2012_case_1 import plotter
 from PySDM.simulation.physics.constants import si
